my-service/pyproject.toml
[tool.poetry]
name = "my-service"
version = "0.1.0"
description = ""
authors = ["shahafido <shahaf@fidocredit.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
pydantic = "^2.5.3"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
aiohttp = "^3.11.11"
asyncio = "^3.4.3"
requests = "^2.32.3"
jwt = "^1.3.1"
cachetools = "^5.5.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
